Fix 1.66 clippy warnings

+++ vhdl_ls/src/vhdl_server.rs
@@ -456,8 +456,8 @@ fn srcpos_to_location(pos: &SrcPos) -> Location {
 
 fn from_lsp_pos(position: lsp_types::Position) -> vhdl_lang::Position {
     vhdl_lang::Position {
-        line: position.line as u32,
-        character: position.character as u32,
+        line: position.line,
+        character: position.character,
     }
 }
